The Virtual Dictionary is an online dictionary that we created as the result of a school lesson plan for our Shipbuilding and Ship Repair class. Through the website, people can read the articles themselves, as it files the uploaded terms into the virtual dictionary. This project provides transportation services in Turkish and English to meet the needs of people who wish to improve their knowledge about shipbuilding.
The Virtual Dictionary is an online dictionary that we created as the result of a school lesson plan for our Shipbuilding and Ship Repair class. Through the website, people can read the articles themselves, as it files the uploaded terms into the virtual dictionary. This project provides transportation services in Turkish and English to meet the needs of people who wish to improve their knowledge about shipbuilding.

Collaborated as a team to put the meanings of these words in the dictionary, I developed a program nicknamed "Adminpanel" to avoid confusion when putting words together, to prevent students from entering the same words, to see which classmate added which word, to ensure that the words added to the dictionary are alphabetical, to enter more than one word at the same time. Thanks to this program, an environment is prepared for solving all these and similar problems and working together.

Only students registered with the White List can enter the "Adminpanel" program and enter words by entering their email address and passwords that have been designated for them by our school. My program that I wrote works in conjunction with Microsoft SharePoint and it automatically generates the students' word entries in the sharing site database and saves them to the "dict_data" spreadsheet created in the SharePoint platform, so we create the database, and it also provides the possibility for common use. Besides these, My Program saves every word that students enter, registration date and time, student emails in SharePoint to check the entries made by students in the form of ".txt" file. An example record format is "ender.miriz@pru.edu.tr 28.05.2022 22 30.txt."

I added the user manual of my program named "Adminpanel" to the folder as ".pdf". You can see the parts you have in mind there.